Make ses-range environment agnostic.
* lisp/ses.el (ses-sym-rowcol): Do not make border effect on interned
symbols.
(ses-formula-references): Do not use ses-range implementation any
longer, as the ses-range is now environnement agnostic, and as such do
not expand any longer to a list of symbols.
(ses-range): Do not make border effect on interned symbols, and be
environment agnostic. As such ses-range macro now expand to a sexp the
evaluation of which yields a list of symbol, and no longer to a list
of symbol with some function call to make it evaluate to the
list of values.

* test/lisp/ses-tests.el (ses-expand-range): New test.